split sis reducer into balances and courses
