+++ source/flux/parts/balances.js
@@ -2,17 +2,9 @@
 
 import {type ReduxState} from '../index'
 import {getBalances, type BalancesShapeType} from '../../lib/financials'
-import {
-	loadCachedCourses,
-	updateStoredCourses,
-	areAnyTermsCached,
-} from '../../lib/course-search'
-import type {CourseType} from '../../lib/course-search'
 
 const UPDATE_BALANCES_SUCCESS = 'sis/UPDATE_BALANCES_SUCCESS'
 const UPDATE_BALANCES_FAILURE = 'sis/UPDATE_BALANCES_FAILURE'
-const LOAD_CACHED_COURSES = 'sis/LOAD_CACHED_COURSES'
-const COURSES_LOADED = 'sis/COURSES_LOADED'
 
 type UpdateBalancesSuccessAction = {|
 	type: 'sis/UPDATE_BALANCES_SUCCESS',
@@ -47,51 +39,7 @@ export function updateBalances(
 	}
 }
 
-type LoadCachedCoursesAction = {|
-	type: 'sis/LOAD_CACHED_COURSES',
-	payload: Array<CourseType>,
-|}
-type CoursesLoadedAction = {|
-	type: 'sis/COURSES_LOADED',
-|}
-
-export type LoadCourseDataActionType = ThunkAction<
-	LoadCachedCoursesAction | CoursesLoadedAction,
->
-export type UpdateCourseDataActionType = ThunkAction<
-	LoadCachedCoursesAction | CoursesLoadedAction,
->
-
-export function loadCourseDataIntoMemory(): LoadCourseDataActionType {
-	return async dispatch => {
-		const areAnyCached = await areAnyTermsCached()
-
-		if (!areAnyCached) {
-			return
-		}
-
-		const cachedCourses = await loadCachedCourses()
-		dispatch({type: LOAD_CACHED_COURSES, payload: cachedCourses})
-		dispatch({type: COURSES_LOADED})
-	}
-}
-
-export function updateCourseData(): UpdateCourseDataActionType {
-	return async dispatch => {
-		const updateNeeded = await updateStoredCourses()
-
-		if (updateNeeded) {
-			const cachedCourses = await loadCachedCourses()
-			dispatch({type: LOAD_CACHED_COURSES, payload: cachedCourses})
-			dispatch({type: COURSES_LOADED})
-		}
-	}
-}
-
-type Action =
-	| UpdateBalancesActions
-	| LoadCachedCoursesAction
-	| CoursesLoadedAction
+type Action = UpdateBalancesActions
 
 export type State = {|
 	balancesErrorMessage: ?string,
@@ -101,9 +49,6 @@ export type State = {|
 	mealsRemainingToday: ?string,
 	mealsRemainingThisWeek: ?string,
 	mealPlanDescription: ?string,
-	allCourses: Array<CourseType>,
-	courseDataState: 'not-loaded' | 'ready',
-	validGEs: string[],
 |}
 
 const initialState = {
@@ -114,15 +59,13 @@ const initialState = {
 	mealsRemainingToday: null,
 	mealsRemainingThisWeek: null,
 	mealPlanDescription: null,
-	allCourses: [],
-	courseDataState: 'not-loaded',
-	validGEs: [],
 }
 
-export function sis(state: State = initialState, action: Action) {
+export function balances(state: State = initialState, action: Action) {
 	switch (action.type) {
 		case UPDATE_BALANCES_FAILURE:
 			return {...state, balancesErrorMessage: action.payload.message}
+
 		case UPDATE_BALANCES_SUCCESS: {
 			const p = action.payload
 			return {
@@ -136,11 +79,6 @@ export function sis(state: State = initialState, action: Action) {
 				balancesErrorMessage: null,
 			}
 		}
-
-		case LOAD_CACHED_COURSES:
-			return {...state, allCourses: action.payload}
-		case COURSES_LOADED:
-			return {...state, courseDataState: 'ready'}
 
 		default:
 			return state
